feat: Refactor & Add Layla Conversation Analyzer modules
- Implemented analytics module with dashboard and key metrics visualization.
- Created authentication module for secure user access.
- Developed chat explorer functionality for browsing conversations.
- Added data loading functions for cloud and local CSV sources.
- Introduced keyword search feature for message filtering across conversations.
- Implemented utility functions for Arabic text detection and translation.
