+++ modules/__init__.py
@@ -1,0 +1,3 @@
+"""
+Layla Conversation Analyzer - Module package initialization
+"""
